+++ structurizr-core/src/main/java/com/structurizr/view/ColorScheme.java
@@ -1,0 +1,11 @@
+package com.structurizr.view;
+
+/**
+ * Represents light or dark mode color schemes.
+ */
+public enum ColorScheme {
+
+    Light,
+    Dark
+
+}